Split the `sockopt` modules into their own files.
